Add Constraints to Product Models and Validations to Product Controller

+++ Controllers/ProductsController.cs
@@ -23,31 +23,81 @@ namespace FullStackDev.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Product>>> Get()
         {
-            return await _productsService.GetProducts();
+            var products = await _productsService.GetProducts();
+            return Ok(products);
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<Product>> Get(string id)
         {
-            return await _productsService.GetProduct(id);
+            var product = await _productsService.GetProduct(id);
+            return Ok(product);
         }
 
         [HttpPost]
         public async Task<ActionResult<Product>> Post([FromBody] Product product)
         {
-            return await _productsService.PostProduct(product);
+            var newProduct = new Product();
+            try
+            {
+                newProduct = await _productsService.PostProduct(product);
+            }
+            catch (Exception e)
+            {
+                ModelState.AddModelError("PostProduct Error", e.Message);
+                return BadRequest(ModelState);
+            }
+            return Created(new Uri($"{Request.Path}/{newProduct.Id}", UriKind.Relative), newProduct);
         }
 
         [HttpPut("{id}")]
         public async Task<ActionResult<Product>> Put(string id, [FromBody] Product product)
         {
-            return await _productsService.PutProduct(id, product);
+            if (id != product.Id)
+            {
+                ModelState.AddModelError("PutProduct Error", "Mismatched Ids");
+                return BadRequest(ModelState);
+            }
+
+            try
+            {
+                var prod = await _productsService.GetProduct(id);
+                if (prod == null)
+                {
+                    ModelState.AddModelError("PutProduct Error", "Product does not exist");
+                    return BadRequest(ModelState);
+                }
+                await _productsService.PutProduct(id, product);
+            }
+            catch (Exception e)
+            {
+                ModelState.AddModelError("PutProduct Error", e.Message);
+                return BadRequest(ModelState);
+            }
+            return Accepted(product);
         }
 
         [HttpDelete("{id}")]
-        public async Task Delete(string id)
+        public async Task<ActionResult> Delete(string id)
         {
-            await _productsService.DeleteProduct(id);
+            try
+            {
+                var prod = await _productsService.GetProduct(id);
+                if (prod == null)
+                {
+                    ModelState.AddModelError("DeleteProduct Error", "Product does not exist");
+                    return BadRequest(ModelState);
+                }
+
+                await _productsService.DeleteProduct(id);
+            }
+            catch (Exception e)
+            {
+                ModelState.AddModelError("DeleteProduct Error", e.Message);
+                return BadRequest(ModelState);
+            }
+
+            return NoContent();
         }
     }
 }
